enh(cli): improve handlind of term width shrink
